Changed remaining Location references in DataDriver to ExposureLocation.  And refactored modules into respective packages
