+++ app/src/main/java/io.github/getExposure/database/ExposureLocation.java
@@ -1,4 +1,4 @@
-package exposure.exposure;
+package io.github.getExposure.database;
 
 import java.util.*;
