+++ app/src/main/res/layout/activity_profile_view.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfileActivity"
+    tools:context=".profile.ProfileActivity"
     android:paddingTop="10dp">
 
     <TextView
