+++ app/src/main/res/layout/fragment_login.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginFragment">
+    tools:context=".profile.LoginFragment">
 
     <TextView
         android:layout_width="336dp"
