+++ app/src/test/java/exposure/exposure/LocationTest.java
@@ -1,4 +1,4 @@
-package exposure.exposure;
+package io.github.getExposure.database;
 
 import org.junit.Test;
